\name{HRTools-package}
\alias{HRTools-package}
\alias{HRTools}
\docType{package}
\title{Convenience functions for \code{adehabitatHR}}
\description{A set of convenience functions built on top of the \code{adehabitatHR} package, to speed up batch home range processing.}
\details{
\tabular{ll}{
Package: \tab HRTools\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-19\cr
License: \tab GPL\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Damiano G. Preatoni and Francesco Bisi

Maintainer: Damiano G. Preatoni <prea@uninsubria.it>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
